app/requirements.txt
openai
streamlit
load_dotenv